Move method to only class it's used in
