+++ collect_app/src/main/java/org/odk/collect/android/utilities/IconUtils.java
@@ -21,10 +21,8 @@ import android.graphics.Bitmap;
 import android.graphics.Canvas;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
-import androidx.core.content.ContextCompat;
 
-import org.odk.collect.android.R;
-import org.odk.collect.forms.instances.Instance;
+import androidx.core.content.ContextCompat;
 
 public final class IconUtils {
 
@@ -50,20 +48,5 @@ public final class IconUtils {
         drawable.setBounds(0, 0, canvas.getWidth(), canvas.getHeight());
         drawable.draw(canvas);
         return bitmap;
-    }
-
-    public static int getSubmissionSummaryStatusIcon(String instanceStatus) {
-        switch (instanceStatus) {
-            case Instance.STATUS_INCOMPLETE:
-                return R.drawable.form_state_saved;
-            case Instance.STATUS_COMPLETE:
-                return R.drawable.form_state_finalized;
-            case Instance.STATUS_SUBMITTED:
-                return R.drawable.form_state_submited;
-            case Instance.STATUS_SUBMISSION_FAILED:
-                return R.drawable.form_state_submission_failed;
-        }
-
-        throw new IllegalArgumentException();
     }
 }
